+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Enhancements:
 
+- feat(compute/init): Offer all available starter kits for each language. ([#1846](https://github.com/fastly/cli/pull/1846))
+
 ### Dependencies:
 - build(deps): `github.com/klauspost/compress` from 1.19.1 to 1.19.2 ([#1881](https://github.com/fastly/cli/pull/1881))
 - build(deps): `github.com/pierrec/lz4/v4` from 4.1.27 to 4.1.28 ([#1881](https://github.com/fastly/cli/pull/1881))